new logic moved all post game code from canvas comp into separate service
