+++ threejs-ng-snake/src/app/canvas-comp/canvas-comp.component.ts
@@ -1,7 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 // todo new logic refresh
-import { Router } from '@angular/router';
-import { Location } from '@angular/common';
 import * as THREE from 'three';
 import { OrbitControls } from 'three/examples/jsm/controls/OrbitControls';
 import { FontLoader } from 'three/src/loaders/FontLoader';
@@ -11,6 +9,7 @@ import { ObjBuilderService } from '../services/obj-builder.service'
 import { SceneHelperService } from '../services/scene-helper.service'
 import { FontBuilderService } from '../services/font-builder.service'
 import { ScoreboardHelperService } from '../services/scoreboard-helper.service'
+import { PostGameHelperService } from '../services/post-game-helper.service'
 
 import { Stats } from '../js/stats'
 
@@ -75,7 +74,7 @@ export class CanvasCompComponent implements OnInit {
     private lastKeyRefresh = 0
 
     // todo new logic game stop time
-    private gameStopTime = 0
+//     private gameStopTime = 0
 
     //
 //     private sceneService: any = undefined;
@@ -85,9 +84,7 @@ export class CanvasCompComponent implements OnInit {
                 private sceneService: SceneHelperService,
                 private fontService: FontBuilderService,
                 private scoreboardService: ScoreboardHelperService,
-                // new logic refresh
-                private router: Router,
-                private location: Location
+                private postGameHelper: PostGameHelperService,
                 ) {
 
         this.scene = new THREE.Scene();
@@ -182,17 +179,13 @@ export class CanvasCompComponent implements OnInit {
                         child.userData.deleteText()
                     })
                     timerGroupObj.children = []
-    //                 this.fontService.addFont("Time's up!!", this.scene, environment.timeWordGroupName, environment.timerGroupPos)
                     this.timerElapsed+=1
 
                 }
             }
 
         }
-        // todo new logic refresh keys
-//         if((timestamp-this.lastKeyRefresh) > environment.keyRefreshRate){
-        // todo new logic only need two mdoes for refresh wireframe
-//         if((environment.postGameMode == "" || environment.postGameMode == environment.modeName3 || environment.postGameMode == environment.modeName4) && (timestamp-this.lastKeyRefresh > environment.keyRefreshRate)){
+
         // todo new logic exclude game mode 3 gameplay else refresh
         if(environment.postGameMode != environment.modeName2 && (timestamp-this.lastKeyRefresh) > environment.keyRefreshRate){
             this.lastKeyRefresh = timestamp
@@ -207,233 +200,17 @@ export class CanvasCompComponent implements OnInit {
             }
         }
 
-        if(environment.postGameMode != ""){
-            if(this.gameStopTime == 0){
-                this.gameStopTime = timestamp
-            }
-            // timesUp mode
-            if(environment.postGameMode == environment.modeName1){
-                let timerGroupObj = this.scene.getObjectByName(environment.timeWordGroupName)
-                if(timerGroupObj != undefined){
-                    this.fontService.addFont("Time's up!!", this.scene, environment.timeWordGroupName, environment.timerGroupPos, environment.largeFontSize)
-                }
-                // Entry mode
-                environment.postGameMode = environment.modeName2
-            }
-
-            // todo new logic environment mode
-//             else if(timestamp - this.gameStopTime > 2000){
-            if(timestamp - this.gameStopTime > 2000){
-
-                if(environment.postGameMode == environment.modeName2){
-                    let timerGroupObj = this.scene.getObjectByName(environment.timeWordGroupName)
-                    if(timerGroupObj != undefined){
-                        timerGroupObj.children.forEach((child:any, i:number)=>{
-
-                            child.userData.deleteText()
-                            // @ts-ignore
-                            // todo logic avoid whole block
-                            this.gameStopTime = -1
-                        })
-                        // todo logic add enter name group
-                        this.fontService.addFont(environment.nameEntryString, this.scene, environment.timeWordGroupName, environment.timerGroupPos, environment.largeFontSize)
-                        // create keyboard
-                        let maxChars = Math.floor(environment.keysAlphabet.length/4)
-                        let curX = environment.buttonGroupPos.x + (maxChars*.25*environment.buttonGroupPos.x)
-                        let curY = environment.buttonGroupPos.y
-
-                        // todo new logic add play again button
-                        // todo new logic using env var not "PLAY AGAIN" hardcode
-                        this.fontService.addFont(environment.playAgainString, this.scene, environment.buttonGroupName, new THREE.Vector3(environment.timerGroupPos.x - environment.smallFontSize*7, environment.timerGroupPos.y + environment.smallFontSize*2, environment.buttonGroupPos.z*.85), environment.xSmallFontSize*.80)
-
-                        environment.keysAlphabet.forEach((characterVal:string, index:any) => {
-                            if(index > 0 && index % maxChars == 0){
-                                curX = environment.buttonGroupPos.x + (maxChars*.25*environment.buttonGroupPos.x)
-                                curY -= environment.xSmallFontSize * 2.5
-                            }
-                            this.fontService.addFont(characterVal, this.scene, environment.buttonGroupName, new THREE.Vector3(environment.buttonGroupPos.x+curX, environment.buttonGroupPos.y+curY, environment.buttonGroupPos.z), environment.xSmallFontSize)
-                            curX += environment.xSmallFontSize * 2.5
-
-                        })
-                        // todo new enter button logic
-                        curY -= environment.xSmallFontSize * 2.5
-                        curX/=2
-                        // todo new logic user env var not "ENTER" hardcode
-                        this.fontService.addFont(environment.enterString, this.scene, environment.buttonGroupName, new THREE.Vector3(environment.buttonGroupPos.x+curX, environment.buttonGroupPos.y+curY, environment.buttonGroupPos.z), environment.xSmallFontSize)
-                        this.fontService.addFont(environment.deleteString, this.scene, environment.buttonGroupName, new THREE.Vector3(environment.buttonGroupPos.x-curX, environment.buttonGroupPos.y+curY, environment.buttonGroupPos.z), environment.xSmallFontSize)
-
-                        // mode 3 scoreboard
-                        environment.postGameMode = environment.modeName3
-                    }
-                }
-                if(environment.postGameMode == environment.modeName3){
-                    // todo new logic check keyboard collide
-                    this.builderService.checkLaserKeyboardCollisions(this.scene)
-                    //
-                    let timerGroupObj = this.scene.getObjectByName(environment.timeWordGroupName)
-                    if(timerGroupObj != undefined){
-                        // todo logic only refresh these if necessary
-                        timerGroupObj.children.forEach((child:any, i:number)=>{
-                            // todo new logic check if message beginning == NAME:
-                            if(child.userData.message != undefined && child.userData.message.substr(0, 5) == environment.nameEntryString.substr(0, 5) && child.userData.message.slice(6, child.userData.message.length) != environment.currWordEntry){
-                                if(child.userData.deleteText != undefined){
-                                    child.userData.deleteText()
-                                }
-                                // @ts-ignore
-//                                 timerGroupObj.children.splice(i, 1)
-                                this.fontService.addFont(environment.nameEntryString + environment.currWordEntry, this.scene, environment.timeWordGroupName, environment.timerGroupPos, environment.largeFontSize)
-                                return
-                            }
-                        })
-
-                    }
-
-                    //
-
-                }
-
-                if(environment.postGameMode == environment.modeName4){
-                    let buttonGroup = this.scene.getObjectByName(environment.buttonGroupName)
-                    if(buttonGroup != undefined && buttonGroup.children.length!=0 && environment.scoreboardObject[0] != 2){
-                        buttonGroup.children.forEach( (child:any, i:number) => {
-                            if(child.userData.deleteText != undefined){
-                                child.userData.deleteText()
-                            }
-                        })
-                        // todo make new function for get scoreboard api
-                    }
-                    // todo here temporary logic might not want to use this method of first element scoreboard
-                    // todo also added logic make sure button group cleared before switching, weird bug with empty name enter causing O and N keys to remain
-                    //@ts-ignore
-                    if(environment.scoreboardObject[0] == -1 && buttonGroup.children.length == 0){
-                        // posting score
-                        this.scoreboardService.postScoreHelper(environment.currWordEntry, environment.userScore)
-                        environment.scoreboardObject = [-2]
-                    }else if(environment.scoreboardObject[0] == -2){
-                        // getting scoreboard
-                        this.scoreboardService.getScoreBoardHelper()
-                        let timerGroupObj = this.scene.getObjectByName(environment.timeWordGroupName)
-                        if(timerGroupObj != undefined){
-                            timerGroupObj.children.forEach((child:any, i:number)=>{
-                                if(child.userData.deleteText != undefined){
-                                    child.userData.deleteText()
-                                }
-                            })
-                        }
-                        let scoreGroup = this.scene.getObjectByName(environment.scoreGroupName)
-                        if(scoreGroup != undefined){
-                            scoreGroup.children.forEach((child:any, i:number) => {
-                                if(child.userData.deleteText != undefined){
-                                    child.userData.deleteText()
-                                }
-                            })
-                        }
-                    }else if(environment.scoreboardObject[0] == 1){
-                        // this: scoreboard object [0] == 1, displaying scoreboard
-                        let timerGroupObj = this.scene.getObjectByName(environment.timeWordGroupName)
-                        if(timerGroupObj != undefined){
-                            if(timerGroupObj.children.length != 0){
-                                timerGroupObj.children.forEach((child:any, i:number)=>{
-                                    if(child.userData.deleteText != undefined){
-                                        child.userData.deleteText()
-                                    }
-                                })
-                            }
-                        }
-                        // todo new logic only put in high score if length 0
-                        //@ts-ignore
-                        if(timerGroupObj.children.length == 0){
-                            environment.scoreboardObject[0] = 2
-
-                        }
-
-                    // todo new logic
-                    // block after here: scoreboard object 0 == 2, displaying scoreboard
-                    }else if(environment.scoreboardObject[0] == 2){
-                        if(environment.timeStampDisplay == -1){
-                            // - 2000 to display faster
-                            environment.timeStampDisplay = timestamp - 2500
-                            // todo add msg "HIGH SCORES" using environment var not hard code
-                            this.fontService.addFont(environment.highScoresString, this.scene, environment.timeWordGroupName, environment.timerGroupPos, environment.largeFontSize)
-                            // todo add msg "PLAY AGAIN" using environment var not hard code
-                            this.fontService.addFont(environment.playAgainString, this.scene, environment.buttonGroupName, new THREE.Vector3(environment.timerGroupPos.x - environment.smallFontSize*7, environment.timerGroupPos.y + environment.smallFontSize*2, environment.buttonGroupPos.z*.85), environment.xSmallFontSize*.80)
-                        }
-
-//                         let scoresList = environment.scoreboardObject[1]
-                        //@ts-ignore
-                        if(environment.scoreStartIndex < environment.scoreboardObject[1].length && timestamp - environment.timeStampDisplay > 3000){
-                            let timerGroupObj = this.scene.getObjectByName(environment.timeWordGroupName)
-                            if(timerGroupObj != undefined){
-                                if(timerGroupObj.children.length != 0){
-                                    timerGroupObj.children.forEach((child:any, i:number)=>{
-                                        // todo new logic avoid high score string update
-                                        if(child.userData.deleteText != undefined && child.userData.message != environment.highScoresString){
-                                            child.userData.deleteText()
-                                        }
-                                    })
-                                }
-                            }
-                            // todo new logic only put in high score if length 0
-                            //@ts-ignore
-                            // todo add msg "HIGH SCORES" using environment var not hard code
-                            let curY = environment.timerGroupPos.y
-                            curY -= environment.largeFontSize*2
-                            let scoresList = environment.scoreboardObject[1]
-                            // todo new logic try to avoid not deleting, cant check if == 0 because high scores object with 2 objects in children list
-                            //@ts-ignore
-                            if(timerGroupObj.children.length <= 2){
-                                //@ts-ignore
-                                scoresList.slice(environment.scoreStartIndex, environment.scoreStartIndex+environment.scoreSliceAmt).forEach((scoreInfo: Array<any>, i:number) => {
-                                    const nameVal = scoreInfo[1]
-                                    const scoreVal = scoreInfo[2]
-                                    // todo new logic incorporate score start index
-                                    const scoreMsg = String(i+1+environment.scoreStartIndex) + " " + nameVal + ":    " + scoreVal
-                                    curY -= environment.smallFontSize * 2
-                                    this.fontService.addFont(scoreMsg, this.scene, environment.timeWordGroupName, new THREE.Vector3(environment.timerGroupPos.x, environment.timerGroupPos.y+curY, environment.timerGroupPos.z), environment.smallFontSize)
-                                })
-                                // new logic time of displaying last scores
-                                environment.timeStampDisplay = timestamp
-                                // new logic update scoreStartIndex
-                                environment.scoreStartIndex += environment.scoreSliceAmt
-                                // @ts-ignore
-                                if(environment.scoreStartIndex >= environment.scoreboardObject[1].length){
-                                    environment.scoreStartIndex = 0
-                                }
-                            }
-
-                        }
-                        this.builderService.checkLaserKeyboardCollisions(this.scene)
-
-                    }else if (environment.scoreboardObject[0] == 3){
-                        this.refreshPagePlayAgain()
-                    }
-
-                }
-            }
-
-
-        }
-
         let scoreGroup = this.scene.getObjectByName(environment.scoreGroupName)
         if(scoreGroup != undefined && this.userScorePrev != environment.userScore && environment.gameStart == true){
             this.userScorePrev = environment.userScore
-//             scoreGroup.children.forEach((child:any) => {
-//                 child.userData.deleteText()
-//             })
-//             todo new logic splice
             scoreGroup.children.forEach((child:any, i:number) => {
                 if(child.userData.deleteText != undefined){
                     child.userData.deleteText()
                 }
-                //@ts-ignore
-//                 scoreGroup.children.splice(i, 1)
             })
             scoreGroup.children = []
             this.fontService.addFont(String(environment.userScore), this.scene, environment.scoreGroupName, environment.scoreGroupPos, environment.largeFontSize)
         }
-
-//         this.secondsElapsed = timestamp - this.lastSecondStart
-
 
         // logic arrow helpers
         if(this.cameraHelpers == true){
@@ -445,8 +222,6 @@ export class CanvasCompComponent implements OnInit {
         }
 
         // main logic asteroids
-        // todo move this to obj service, use object methods
-        // todo this just test helper for movement
         if(timestamp-this.lastRotationStart > environment.rotationFramerate){
             this.lastRotationStart = timestamp
             // spin each asteroid and rotate around center y axis
@@ -464,40 +239,26 @@ export class CanvasCompComponent implements OnInit {
             });
         }
 
+        if(environment.postGameMode != ""){
+            this.postGameHelper.postGameRouter(this.scene, timestamp, this.builderService, this.scoreboardService, this.fontService)
+        }
 
         // update laser: init new, set depleted, check max distance and delete
         this.sceneService.updateLaser(this.scene, controlsTarget)
-//         if(environment.gameStart == true){
-//         this.builderService.checkLaserCollisions(this.shapesArray, this.scene);
-            // todo new logic box helpers bool
-            this.builderService.checkLaserCollisions(this.shapesArray, this.scene, this.boxHelpers);
-//         }
+        this.builderService.checkLaserCollisions(this.shapesArray, this.scene, this.boxHelpers);
         this.render_all();
         this.stats.update();
         requestAnimationFrame(this.animate);
-//         this.last = timestamp;
-//         console.log(typeof timestamp)
     }
 
-    refreshPagePlayAgain() : void{
-//         this.router.navigateByUrl("/refresh", { skipLocationChange: true }).then(() => {
-//             console.log(decodeURI(this.location.path()))
-//             this.router.navigate([decodeURI(this.location.path())])
+//     getWordApi() : void {
+//         this.wordService.getWord().subscribe(data => {
+//             let jsonPickleStr = JSON.stringify(data);
+//             let jsonPickle = JSON.parse(jsonPickleStr);
+//             let pickleWord = jsonPickle.pickle_time
+//             this.wordGet = pickleWord
 //         })
-//         location.reload()
-        window.location.reload()
-        // must set scoreboard object to avoid looping and reloading multiple times
-        environment.scoreboardObject[0] = 4
-    }
-
-    getWordApi() : void {
-        this.wordService.getWord().subscribe(data => {
-            let jsonPickleStr = JSON.stringify(data);
-            let jsonPickle = JSON.parse(jsonPickleStr);
-            let pickleWord = jsonPickle.pickle_time
-            this.wordGet = pickleWord
-        })
-    }
+//     }
 
     // todo new logic post word API
