+++ threejs-ng-snake/src/app/services/post-game-helper.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { PostGameHelperService } from './post-game-helper.service';
+
+describe('PostGameHelperService', () => {
+  let service: PostGameHelperService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(PostGameHelperService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
